[Typ] Create method to get the sub directory for artifacts
This CL moves the logic that derives the sub directory where the
artifacts will be written to for a specific run into its own method.
It will also modify typ so that it stores the full relative path of
artifacts to the artifacts dictionary. This CL also adds an optional
constructor argument which can be used to allow overwriting artifacts
when a test is repeated within a single iteration.

TBR=dpranke@chromium.org

Bug: chromium:982208
Change-Id: I3df46ebde83311de9b91cdb1e3baff75bb275dcc
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1837272
Commit-Queue: Rakib Hasan <rmhasan@google.com>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>