Fix results2 dependency on internal functions
It turns out results2.py depended on functions meant to be internal-only
in the read_value quest. Since we've refactored those names to not have
the leading underscore, we started failing in production for getting the
results of jobs attempting to generate the new version of the results
viewer.

Also, it turns out, we didn't catch this because there are no unit tests
covering this part of the results2 implementation.

R=abennetts@google.com, simonhatch@chromium.org
TBR=abennetts@google.com,simonhatch@chromium.org
Bug: chromium:1012187
Change-Id: Ie5e33d6197f410f95eac3cb594253a2192a21b5e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1846585
Reviewed-by: Dean Berris <dberris@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>