[Telemetry] Simplify Telemetry progress reporter
To make it more consistent with how test results are reported elswhere,
we directly use the story_run.test_path instead of trying to build its
own identifier.

The only visible effect is for the case of stories with grouping keys
which used to be written by the progress reporter, but wont be after
this change.

Grouping keys in the progress report were added long time ago by
https://chromium.googlesource.com/catapult.git/+/d2de9db0fccf47ae033b793844a223afefccee82
to help distinguish stories with the same name but different grouping keys.

However, story sets are no longer allowed to have the duplicate stories with
the same name any more, c.f.:
https://cs.chromium.org/chromium/src/third_party/catapult/telemetry/telemetry/story/story_set.py?rcl=9d0252567353a4588b57f6ab775095094bc4d17f&l=99

This CL also removes a call to results.IterStoryRuns which we intend to
deprecate.

Bug: chromium:973837
Change-Id: Idb10fbf5476ecb0fc1458bc3d6a38c87ea9e9190
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1920123
Reviewed-by: Mikhail Khokhlov <khokhlov@google.com>
Commit-Queue: Juan Antonio Navarro Pérez <perezju@chromium.org>