Pinpoint: Engine-Driven Completion Through Updates
This change rounds out the test case for the execution engine and fixes
a few bugs along the way with the way we're updating job-level details.
In particular, we're ensuring that:

- We introduce a 'done' property to the Job model that's transactionally
updated to ensure correctness across multiple concurrent task update
handlers working on the same job.

- When performing a bisection, we only mark the job done when the
execution graph represents the done state (i.e. when the culprit finding
vertex in the graph is marked either failed or completed).

- Update the way we check whether a job is "completed" by taking the
execution engine and the done property into account.

- Update some action lambdas into standalone types with special
formatting for easier debugging through logs provided by the execution
engine core.

At the heart of this change is a bug in the implementation of how we're
determining when a bisection is "done", which we were only able to test
by simulating updates through the task update handling. Unit tests did
not catch this situation because we mocked the effect of other
evaluators, which with a fairly representative end-to-end integration
test (still mocking calls to external services) allowed us to observe
more intermediary states of the job as it evolves. With this change, we
have the beginnings of a way of end-to-end testing the engine through
the task update handler.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1032880
Change-Id: I1681007d08b9a2359474ec2bc4c71684cac4d8b6
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1962130
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>