Use all known changes for determining order
In the Pinpoint Execution Engine, we have access to all of the changes
we could be bisecting through, and can have states that might not have
data yet. In order to show this relative ordering, we need to update the
way we're processing the order of changes by relative order of the
changes we know about in the process of bisection, when serialising this
data.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1031433
Change-Id: I5f29f3f8faf465d9ae6b35af2b338da595493af7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1958294
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>