Add result values to serialized output
This change adds the result_values to the serialized state, to allow the
UI to show information on changes. Even if the list if empty, we should
be providing this information to allow the UI to render some of the
state data regarding the changes we're working on.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034369
Change-Id: I2b41de19c19b48ecff582ecef6ee2b91758004a1
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1965880
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>