Always serialize the find_isolate status
This change ensures that even if the task is pending, that we have
information about the find_isolate part of the job. This means if we got
the builds from cached information, that it should show up immediately.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1034369
Change-Id: I88e4decd6ae6d3800aafdba632a17529698fb3a9
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1965881
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>