Dashboard: Integrate an approximation of E-Divisive
This change implements an approximation of the E-Divisive algorithm for
culprit finding in timeseries data. It's an approximation because we
are only finding a single change point, as opposed to finding all the
change points we can see.

For now, we're writing out log messages so that we can watch the number
of differences we're seeing in production so that we can quantify the
potential for false negatives and false positives. Once we have more
data about how well this alternative implementation is doing, then we
can make a data-driven decision in changing the default algorithm we're
using.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:922256
Change-Id: I020781c9551cc7be472d5127be3c8a1ff13b63fb
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1960046
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>