Dashboard: Tweak Clustering For Better Placement
This change moves the potential change point in clustering to the
"right" cluster to allow us to prefer "newer" change-points as potential
culprits and have a better placement of the change-point we're
attempting to find. This works because when we're making the decision to
recurse into the clusters to refine our placement, having the
most-likely change be on the first index of the cluster means it's in
the margins of the "minimum segment size". For us to reject this
potential change point, we'd need to have found another significantly
different potential change point within the cluster including the
potential change point.

The heuristic to always assume that newer points (those on the right of
the timeline) are more likely to be culprit points works for the
Dashboard. It also works nicely because we want to choose the point
where the "change" in performance measurements has "stuck", because
points in the timeline refer to ranges.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:922256
Change-Id: I90b9183238335daf2907b0106b253c3fb7c33b7b
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1985499
Commit-Queue: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>