Switch default to sheriff-config
Use the subscriptions (from sheriff-config) stored in Anomaly to
replace sheriff. We are still in the middle of migration, so both config
need to updated and keep synchronized. And because sheriff-config
doesn't provide the list of subscriptions yet, we still need to use old
sheriff (at least their names) for some parts of the UI.

We also remove the email summary feature from the Dashboard which
groups alerts by sheriff configuration. This is a conscious choice in
preparation for auto-filing bugs in Monorail.

Bug: chromium:953181
Change-Id: I591a99848060dd437d0c68814a319c052ce0b2c2
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/1996967
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>