Teaching histograms to deserialize from protos.
For this CL, I thought everything should read the proto as natively
as possible since the proto will be a first-class format (alongside
histogram JSON and the unfinished "serialized" format).

I'm going to generate as many protos as I can from one of our tests
and run through this new code to verify it's correct, but I think
this is a good start at least.

This solution uses an import hack to somewhat gracefully deal with
a missing histogram_pb2.py. This means Telemetry and other
dependencies that don't use protos will keep working without
having to generate the proto stub.

I have to add a protobuf lib to the presubmit python invocations,
and I took the one in third_party/gae_ts_mon/gae_ts_mon/protobuf.
We should consider adding third_party/protobuf to follow the law
of Demeter, but this is OK for now I think. It's just the
presubmit anyway.

Bug: chromium:1029452
Change-Id: Iae3d01fb863e499f2063e4edc85b7f45de8b0eb3
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2027457
Reviewed-by: Dean Berris <dberris@chromium.org>
Commit-Queue: Patrik Höglund <phoglund@chromium.org>