Increase max concurrency for histograms-queue
This allows us to drain the queue quicker with the new limits we have on
the number of instances running. We can afford to queue up here because
this is a background processing job, and we have a higher limit on the
per-task request completion.

R=abennetts@google.com, fancl@chromium.org

Change-Id: I9d85293eb9a801ad4490b3be287a68693783a38e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2050068
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>