[trace viewer] Fix test runner under polyfill
After this CL, tests.html can load and run tests.

This is useful to get a sense of how many features work under the
polyfill. So far it shows there are 86/2510 failures.

This does not break tests that currently run on bots, because we run
tests by turning on webcomponentv0 with a flag.

Bug: chromium:1036492
Change-Id: Id02304c18a4a6a111b9545518450de823445ad97
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2057330
Commit-Queue: Deepanjan Roy <dproy@chromium.org>
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>