[Typ] Set default string arguments to None in Expectation constructor
We need to set the default values of string parameters for the
Expectation constructor to None. Then we need to set the member
variables into empty strings in the constructor. This will ensure
that string member variables of Expectation instances are equal
to each other whenever they are not set.

TBR:dpranke@chromium.org
Bug: chromium:986447
Change-Id: I00d7f7f3d135d38f8b30e1b0643f9c4a875d5273
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2067331
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Rakib Hasan <rmhasan@google.com>