[Telemetry] Look for local minidump_stackwalk
Updates minidump_symbolizer to look for minidump_stackwalk in the
build output directory before falling back to using the dependency
manager. This is so that we use the most appropriate version of the tool
for the version of Chrome we're using if it's available.

Bug: chromium:1040502
Change-Id: Ifee2f43b0f7ac5e9949208d9935b0aa0d287468c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2068128
Reviewed-by: Caleb Rouleau <crouleau@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>