Pinpoint: Link to task isolate on failure
Before this change, we attempted to parse the output log from Swarming
and find exceptions to present to the user. Because we can't rely on
Telemetry providing the information the output, we instead should link
to the output isolate produced by the task.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1055568
Change-Id: Id18b83f197a09302968534bf4a026c5fb2824dc4
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2071755
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>