[systrace] Add android_cgroup_agent
The cgroup information is important to track scheduling behavior of
important apps in Android.  It adds all cgroup events when 'sched'
category is selected but it only covers when a task moves to a
different cgroup or a new cgroups was added or removed.  So there's no
way to know which cgroup existing (non-moving) tasks belong to.

The new cgroup agent will record those initial cgroup information in
the trace.  After that cgroup trace events can track all the changes.
For now, it's only enabled on android target and track cpuset cgroups.

Change-Id: Ic1a5d4320975a16c85bdd5aa4bfbb86f1c1d1315
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2073322
Commit-Queue: Namhyung Kim <namhyung@google.com>
Reviewed-by: John Reck <jreck@google.com>
Reviewed-by: Deep Roy <dproy@chromium.org>