[NetLog Viewer] Show loaded filename in title
When loading a NetLog file into the viewer, the title was shown only
within the body of the page, but the title was left empty. To simplify
working with multiple instances of the NetLog Viewer at one time, this
change will also show the loaded capture's filename in the title.

Bug: chromium:1058919
NOPRESUBMIT=true
Change-Id: I3355dd4d909667ed64c0c4e76bbb271fa81ef5f7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2090391
Commit-Queue: Eric Lawrence [MSFT] <ericlaw@microsoft.com>
Reviewed-by: Eric Roman <eroman@chromium.org>