Dashboard: Terminate on Permutation Test Convergence
In this change we ensure that if permutation testing suggests new bounds
that are the same as the current bounds, that we terminate further
iterations and log an issue. This fixes the edge-cases we're
encountering in production where permutation testing will keep finding
the same bounds without the algorithm detecting the infinite loop.

We also use array.array(...) internally for speeding up the computation
involving numeric values, improving memory locality by keeping the
numbers adjacent in memory.

We also take the opportunity to remove the requirement for a
min_segment_size in the API for the cluster-based detection (E-Divisive
implementation).

R=fancl@chromium.org

Bug: chromium:1067589
Change-Id: I52b6e148658fdf005884f01878be920eea8df23d
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2135550
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>