Dashboard: Add timestamps to anomaly metadata
This change adds the earliest and latest timestamp for the values we're
seeing in the database when creating the anomalies. This is important in
computing 'time-to-detection' where we're getting the difference between
the time an anomaly is detected, and the earliest time for the rows used
in the anomaly detection input.

This is in preparation for adding useful information in the auto-triaged
issues we'll be filing.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1064955
Change-Id: I818ce8bd117e09327d02ab9aa19a0304e5f08253
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2166881
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Reviewed-by: Andrew Bennetts <abennetts@google.com>