Add TestExpectations.is_default_pass
This allows the client know if the test expectation is an explicit
Pass or a default Pass.

Also let [ Slow ] and [ RetryOnFailure ] imply default pass only,
instead of always implicitly adding a [ Pass ] expectation. For example,
[ Linux ] foo/bar [ Failure ]
foo/bar [ Slow ]
we should not treat [ Pass ] as a valid expectation on Linux.
Previously we were unable to mark the test always failing on Linux.

Bug: chromium:1072015
Change-Id: I00e1488ca6a0db9acd497289d511e042752d97b5
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2173206
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Reviewed-by: Rakib Hasan <rmhasan@google.com>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>