auto-triage: Exclude recovered anomalies
This change ensures that we do not consider recovered anomalies as
regressions that we should file issues for. In the future we should also
ensure that recovered alerts are marked in the issues where they occur.
It will also help if we can auto-close issues where all the anomalies in
a group have recovered.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1081607
Change-Id: Iaf92bf42abd2d5f43b9b70e2cc8766c806997341
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2195433
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>