[Telemetry] Standardize build directory usage
Standardizes the usage of build directories by making
ChromeBrowserBackend take a build directory argument in its constructor.

Also updates the way that children find the build directory to be
consistent, i.e. the chosen build directory will always correspond to
the browser that's actually being used. Previously, it was technically
possible for them to differ on Android.

This is a prerequisite for looking for locally built dependencies before
falling back to downloaded ones, as we need to look in the build
directory for those.

Bug: chromium:1026296
Change-Id: I26ddb3a8e06b602f72c6db1404b3e59c3c1b4322
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2203115
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>