pinpoint: Add the "pin" field in bisection form
This allows us to provide a patch (a Gerrit URL) which will be applied
to a bisection job. This shows up in reruns and in new bisection jobs
being started. It mirrors functionality that is available from the
Dashboard UI currently that's been missing from the Pinpoint UI.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1079183
Change-Id: I1002e2b033f86ff28b6bfd79e49308e36fae6d86
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2217992
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>