auto-bisect: Do not bisect ignored and recovered anomalies
This change ensures that the Dashboard doesn't start bisections on
anomalies that are purposely ignored or marked recovered.

We also more completely associate issues to un-triaged anomalies that
are associated with a group (instead of just to regressions).

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1087799
Change-Id: I7397f749ae93d3bd68d35a5ca8d8934ca96a7851
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2217986
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>