[Telemetry] More performance mode control
This CL introduces the --performance-mode command-line flag that allows
user to explicitly choose the desired performance mode for the device.
Also adds an option to pin the benchmark execution to little
cores only.

The default behaviour is left unchanged: the device is set to high
performance mode before the execution of the benchmark, and to normal
mode after it finishes.

Bug: chromium:1096589
Change-Id: I8d10e354a10139dbcfcd280d7d8a09246850833b
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2252182
Commit-Queue: Mikhail Khokhlov <khokhlov@google.com>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>