pinpoint: Break out details of runtime
This change breaks out the time a job was created, how long it's been in
the queue, when it's started, and how long it's been running (or how
long it ran for). We need to change the serialized model to include the
start time for each job, and handle the cases when a job is cancelled.

Unfortunately we don't have JavaScript testing set up for these elements
(yet) so we're going to have to see this in action in production and
follow up on any inconsistencies we find.

R=abennetts@google.com, fancl@chromium.org

Change-Id: Ia22a290a3e37c9e42bcb47506f2f81128c45bcb1
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2266094
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>