exec: Use successive doubling for sample sizes
This change makes Pinpoint double the number of attempts up to a cap of
160 per change starting from 10 attempts when attempting to get stronger
statistical sampling. This means we get up to 4 iterations to get from
10 to 160 attempts with successive doubling.

We also change it so that if we decide to scale attempts for change[a]
that we'll also scale attempts for change[b] if they both have the same
number of attempts. This change should greatly reduce the amount of
waiting that's currently happening when we end up needing to get more
samples to get a stronger statistical signal on whether we should bisect
between two changes.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1101200
Change-Id: Ideb9ee7aa37c6db8c5c13947712e58da3424724f
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2275571
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>