debt: Remove special cases for Pinpoint jobs
Given that all the Chromium Telemetry benchmarks are now all runnable
through a single isolate target, we're removing the special cases in the
Dashboard code for using different kinds of isolate targets. A follow-on
change to the Pinpoint UI should reflect the removal of these special
cases as well.

In this change we also take the opportunity to remove special-case
lookups on revision numbers that have been at least a year old.
Pinpoint is unlikely able to work with those revisions anymore so we
don't bother with the range lookups anymore.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:950165
Change-Id: I5bd16aef905e2ff98607eab0e0a12ed0aa916862
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2277698
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Auto-Submit: Dean Berris <dberris@chromium.org>