metrics: Add diagnostics to console error metric
https://crrev.com/c/2302837 adds extra details (message and url) to
ConsoleError trace events. This CL adds these info as diagnositcs to
the console error metric so the metric is easier to debug. Additionally,
the pid of the slice is added as a diagnostic since that's also quite
useful for debugging.

The code safely skips adding these details if we're using an older
chrome version and the extra trace arguments are not available on the
ConsoleError slice.

Bug: chromium:1086962
Change-Id: I50403d59e812b42ee68afbd62083c6922d104756
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2304259
Reviewed-by: Ulan Degenbaev <ulan@chromium.org>
Commit-Queue: Deep Roy <dproy@chromium.org>