proj-perf-tooling: Update anomaly config schema
This change updates the definition of the anomaly configuration proto to
use the shared `Rules` message to define inclusion and exclusion lists.
It also adds documentation about what the fields in an anomaly
configuration mean, along with default values.

Following changes will make the required changes in the anomaly
detection part of the Dashboard to use the values from the Subscription
proto instead of the values in Datastore.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1110763
Change-Id: I4d76835f4a51773ee79b30d385d46746d4ae83c4
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2325518
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>