Use start & end in the anomaly instead of AlertGroup
We need both test path and revision to retrieve the Row. In some cases,
when AlertGroup includes regressions across different ranges, the Row
may not exist if we use revision in the AlertGroup with test path in
the regression. Always use the revisions in the regression with test
path.

R=abennetts@google.com, dberris@chromium.org

Bug: chromium:1106244
Change-Id: I7036fea40c30ae93f2f5f17573c1b79eabda4fc2
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2326861
Auto-Submit: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>