[Devil] Retry forwarder pkill on -9
Retries pkilling the host forwarder when the first attempt returns a -9
exit code. This seems to indicate that the requested process was alive
when pkill was started, but died sometime between then and pkill
actually sending the terminate signal. Retrying *should* result in a 0
or 1 exit code. If not, then the existing logic will handle that
failure.

Bug: chromium:1038789
Change-Id: I03013a50ac5b89b4819947bf33f3bd0dc030fc02
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2334357
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>