[Telemetry] Add full screenshot support
Adds support for capturing the full contents of a page when capturing a
screenshot instead of just the visible portion. The particular use case
in mind is for running pixel tests on devices with small screens, but
this could be useful elsewhere.

Bug: chromium:1115680
Change-Id: Ibdd2a11a34d7fb24630a5fb212a95a6c46c6b1c0
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2354161
Commit-Queue: John Chen <johnchen@chromium.org>
Reviewed-by: John Chen <johnchen@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>