Improve AdbWrapper#is_emulator to use device codename vs serial
The device serial may not always be the best indicator if a device is an
emulator. For example, tunneling an emulator would appear as
"localhost:port" instead of "emulator-*". This cl changes the check to
use the device codename instead.

Bug: chromium:1124965
Change-Id: Ie5151d20098c6c437f8731eb20e6897574bf8442
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2367333
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Mohamed Heikal <mheikal@chromium.org>
Auto-Submit: Mohamed Heikal <mheikal@chromium.org>