Update /uploads api to return diagnostics data.
Also splitted api into dafault and full info one in order to reduce load on Datastore when information about measurements isn't needed (which is expected to be common).
Also I've used 'diagnostics' instead of 'dimensions' (as in the design doc). Such naming seems to be less confusing.
If the changes in api will be accepted, I'll update the design doc.

Bug: chromium:1122516
Change-Id: Iffc72fa801dbce308af928ae72c0209d8ca69c6e
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2411980
Commit-Queue: Andrey Logvin <landrey@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Reviewed-by: Dean Berris <dberris@chromium.org>