[Devil] Remove forwarder strace
Removes the strace call when retrying the forwarder kill attempt. The
addition of the strace seemed to cause the issue we were trying to
debug to disappear, so remove the strace call to see if the issue starts
occurring again.

Bug: chromium:1038789
Change-Id: If3b1d5126afbf9a1511dbee137bd77391322fb3c
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2427553
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>