Fix dependency_manager path hacking.
This CL changes the path-hacking in dependency_manager/__init__.py
to insert new paths rather than append them.

When you are attempting to manually modify the PYTHONPATH
in a python file in order to be able to import a new package
from some other part of the source tree, it's important to
insert the directory at the front of the path, rather than
appending it to the end of the path.

If you do the latter, then you risk picking up a system-local
version of the module instead, which can result in surprising
problems down the road.

This CL only modifies dependency_manager because I'm trying to
limit the potential scope of changes in order to unblock the
Catapult roll into Chromium. However, someone should probably
audit the rest of the code base for similar bugs.

Bug: chromium:1135502
Change-Id: Ie5f394fd0303be22411555e33660f134369a1e86
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2453332
Reviewed-by: John Chen <johnchen@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>