Add way to have a hidden container slice
I want to group two async slices, one for the time a job is queued, and
another for the time a job is running.  Subslices can achive this, but
in this case I don't want to see the container slice.

Change-Id: Id3c3062f754766e1f06a9aff887475dfa7df4754
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2487761
Auto-Submit: Rob Clark <robdclark@chromium.org>
Reviewed-by: oysteine <oysteine@chromium.org>
Commit-Queue: Rob Clark <robdclark@chromium.org>