[Telemetry] Adds telemetry support to Fuchsia for memory benchmarks.
Adds minimal support for Fuchsia in Telemetry for memory benchmarks.
This starts tracing if it is not already running when the browser starts
(since StartUp tracing is not supported on Fuchsia). Also forces
memory dumping on Fuchsia to be 'light' detailed since 'detailed' is
not supported on Fuchsia.

Change-Id: I8db9ca8da54bd2218cc4125fc879173e344a6d29
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2500418
Commit-Queue: Rohan Pavone <rohpavone@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>