Log trace collect time in TracingBackend.
When running story tab_search:top100:2020,
The trace collect time is enormously long for Windows 10 perf bots,
sometimes leading to timeout.
The issue is not reproducible locally on a Windows 10 machine.
Adding timing to the log for better exposure of the issue.
For more information see the following doc:
https://docs.google.com/spreadsheets/d/1CqM-cKB5z70RTVrEZukWsWB2kOaz6622ZwGpisiKink

Bug: chromium:1152598
Change-Id: I90083a5efb4c173407793722bde54115a5ba3d47
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2593459
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Yuheng Huang <yuhengh@chromium.org>