[telemetry] python3 conversion:unicode_type
Convert telemetry to Python 3 style applying fixers on telemetry/:
  python -m modernize -w -f unicode_type .

The text type is handed already on mock.py and util.py, and this CL did not touch those files.

This is one of the CLs converting Telemetry to be compatible for Python3. The conversion utilizes the modernize module: https://pypi.org/project/modernize/

Bug: chromium:1198408
Change-Id: I208323b5c11cbf07e15f96b95f1dca29a731deb7
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2844610
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>