Only collapses slices that do not have arguments
Trace viewer collapses events with exactly one child which have the exast start and end time, as such events would not add any information to the trace.
The only exception is ehen the trace event has arguments with it that collapsing it makes them unavailable.
This change adds the condition to check for arguments in the slice before collapsing it.

Bug: chromium:1205873
Change-Id: I2fde88be6af911b1ef4d4b47f4e0bcb38da883d6
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2873447
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Behdad Bakhshinategh <behdadb@chromium.org>