pinpoint: Add support for A/B testing flags
This change allows us to specify base and experiment flags that
differ, without requiring a patch to be provided for A/B testing. This
is useful in cases where we'd like to understand the effects of changing
flags for the base and experiment side without requiring different
builds.

The functionality is currently not exposed to the UI, and will instead
be supported through the Pinpoint CLI, which has a more structured API
contract between the gRPC service and the command line interface
utility.

R=abennetts@google.com, fancl@chromium.org, seanmccullough@google.com

Bug: chromium:1201147
Change-Id: I34c3547f8cbf256cb56c2a1bfec16d24940b5566
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2914245
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>