[Pinpoint] Add fallback target
Adds the concept of a fallback target to the FindIsolate quest, which
will be used if the primary target isn't found but the fallback is. This
is an optional argument, so all current uses should be unaffected.

This will be useful for cases where a target changes in Chromium, as we
need to support both the original and new target names for a period of
time.

Bug: chromium:1203798, chromium:1218771
Change-Id: Ib8094d60f86fbfbc69c63720757fa940c476c818
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/2977059
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>