Modified chrome startup tracing agent to work with categories flag.
In order to create a memory profile, the chrome_categories flag passed
to the tools/tracing/profile_chrome_startup script has to be set. The
chrome_categories flag, however, was being ignored because the
chrome_categories flag was not being passed as an argument when the
tracing agent was created. This change modifies
chrome_startup_tracing_agent.py to use the chrome_categories flag.

Bug: chromium:1224717
Change-Id: I30087115eb06b3ca0c3333174ec1f1d0c34d3a47
Reviewed-on: https://chromium-review.googlesource.com/c/catapult/+/3042547
Reviewed-by: ssid <ssid@chromium.org>
Reviewed-by: Deep Roy <dproy@chromium.org>
Commit-Queue: ssid <ssid@chromium.org>